refactor: move mcp consent boundary to api adapter
